refactor: 도메인 패키지 분리
두 가지의 주요 기능을 고려하여 lotto와 winning 패키지를 분리하한다.
- lotto는 로또 구매 및 발행 도메인
- winning은 로또 당첨 결과 산출 도메인
