+++ src/main/java/lotto/LottoGame.java
@@ -3,14 +3,14 @@ package lotto;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
-import lotto.domain.Bonus;
-import lotto.domain.Lotto;
-import lotto.domain.LottoDrawer;
-import lotto.domain.LottoTicket;
-import lotto.domain.NumbersGenerator;
-import lotto.domain.PurchaseAmount;
-import lotto.domain.Rank;
-import lotto.domain.WinningCombination;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoDrawer;
+import lotto.domain.lotto.LottoTicket;
+import lotto.domain.lotto.NumbersGenerator;
+import lotto.domain.lotto.PurchaseAmount;
+import lotto.domain.winning.Bonus;
+import lotto.domain.winning.Rank;
+import lotto.domain.winning.WinningCombination;
 import lotto.external.RandomNumbersGenerator;
 import lotto.view.Input;
 import lotto.view.Output;
