+++ src/main/java/lotto/domain/lotto/PurchaseAmount.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 public class PurchaseAmount {
