+++ src/test/java/lotto/domain/LottoDrawerTest.java
@@ -3,6 +3,10 @@ package lotto.domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.List;
+import lotto.domain.lotto.LottoDrawer;
+import lotto.domain.lotto.LottoTicket;
+import lotto.domain.lotto.NumbersGenerator;
+import lotto.domain.lotto.PurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -16,7 +20,7 @@ class LottoDrawerTest {
         LottoDrawer lottoDrawer = new LottoDrawer(purchaseAmount, fixedNumbersGenerator);
 
         LottoTicket lottoTicket = lottoDrawer.generateLottos();
-        
+
         lottoTicket.getLottos().forEach(lotto ->
                 assertThat(lotto.getNumbers()).isEqualTo(List.of(1, 2, 3, 4, 5, 6))
         );
