+++ src/test/java/lotto/domain/PurchaseAmountTest.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import lotto.domain.lotto.PurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
