+++ src/test/java/lotto/domain/RankTest.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.stream.Stream;
+import lotto.domain.winning.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
