+++ src/test/java/lotto/domain/WinningCombinationTest.java
@@ -6,6 +6,12 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoTicket;
+import lotto.domain.lotto.PurchaseAmount;
+import lotto.domain.winning.Bonus;
+import lotto.domain.winning.Rank;
+import lotto.domain.winning.WinningCombination;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
